Rename Metadata_Editor

+++ lib/services/downloader.dart
@@ -2,8 +2,6 @@ import 'dart:io';
 
 import 'package:youtube_downloader/index.dart';
 import 'package:path/path.dart' as path;
-import 'package:youtube_downloader/services/metadata_editor2.dart';
-import 'package:youtube_downloader/services/metadata_editor3.dart';
 import 'package:youtube_explode_dart/youtube_explode_dart.dart';
 
 
@@ -31,7 +29,17 @@ class Downloader {
     }
 }
 
-
+  String checkTitle(String title) {
+    title = title.replaceAll(r'\', '')
+      .replaceAll('/', '')
+      .replaceAll('*', '')
+      .replaceAll('?', '')
+      .replaceAll('"', '')
+      .replaceAll('<', '')
+      .replaceAll('>', '')
+      .replaceAll('|', '');
+    return title;
+  }
 
   Future<void> downloadVideo(String title, String artist, String album) async {
     print(">> Start Download");
@@ -43,9 +51,6 @@ class Downloader {
     var manifest = await ytExp.videos.streamsClient.getManifest(this.id.toString());
     var video = manifest.video.first;
 
-    // TODO for video here
-    // var video = manifest.video.first;
-
     // Build the directory.
     var dir = await DownloadsPathProvider.downloadsDirectory;
     // create a new Directory to save all Youtube Downloader files
@@ -55,8 +60,10 @@ class Downloader {
         newDir = d;
       });
 
+    title = checkTitle(title);
+
     var filePath = path.join(newDir.uri.toFilePath(),
-        '$title.$videoFileEnding');
+        '$title.${video.container.name.toString()}');
 
     // Open the file to write.
     var file = File(filePath);
@@ -71,18 +78,20 @@ class Downloader {
     await fileStream.flush();
     await fileStream.close();
 
-    // TODO show finished
-    // // Show that the file was downloaded.
-    // await showDialog(
-    //   context: context,
-    //   builder: (context) {
-    //     return AlertDialog(
-    //       content: Text(
-    //           'Download completed and saved to: ${filePath}'),
-    //     );
-    //   },
-    // );
-
+    if (video.container.name.toString() != videoFileEnding) {
+      // Use FFmpeg to convert to .mp4
+      String outputPath = path.join(newDir.uri.toFilePath(),
+        '$title.$videoFileEnding');
+      
+      var command = "-i \"$filePath\" \"$outputPath\"";
+      await FlutterFFmpeg().execute(command)
+        .then((rc) => print("FFmpeg process exited with rc $rc"));
+      
+      file = File(filePath);
+      if (await file.exists()) {
+        await file.delete();
+      }
+    }
 
     // TODO übermehmen aus Musik
     // editMetadata(filePath.toString(), title, artist, album);
@@ -97,13 +106,9 @@ class Downloader {
 
     var video = await ytExp.videos.get(this.id);
   
-
     // Get the streams manifest and the audio track.
     var manifest = await ytExp.videos.streamsClient.getManifest(id);
     var audio = manifest.audioOnly.last;
-
-    // TODO for video here
-    // var video = manifest.video.first;
 
     // Build the directory.
     var dir = await DownloadsPathProvider.downloadsDirectory;
@@ -113,20 +118,10 @@ class Downloader {
         newDir = d;
       });
 
-    title = title.replaceAll(r'\', '')
-      .replaceAll('/', '')
-      .replaceAll('*', '')
-      .replaceAll('?', '')
-      .replaceAll('"', '')
-      .replaceAll('<', '')
-      .replaceAll('>', '')
-      .replaceAll('|', '');
+    title = checkTitle(title);
 
     var filePath = path.join(newDir.uri.toFilePath(),
         '$title.${audio.container.name.toString()}');
-
-    // var filePath = path.join(newDir.uri.toFilePath(),
-    // '$title.$musicFileEnding');
     
     // Open the file to write.
     var file = File(filePath);
@@ -137,77 +132,37 @@ class Downloader {
 
     // TODO Here implementation of a % of download possible
 
-
-    // TODO remove if unused
-    // file = await editMetadata(file, filePath.toString(), title, artist, album);
-
-
-
     // Close the file.
     await fileStream.flush();
     await fileStream.close();
 
-    // TODO show finished
-    // // Show that the file was downloaded.
-    // await showDialog(
-    //   context: context,
-    //   builder: (context) {
-    //     return AlertDialog(
-    //       content: Text(
-    //           'Download completed and saved to: ${filePath}'),
-    //     );
-    //   },
-    // );s
-  
-    // Use FFmpeg to convert to .mp3
-    String outputPath = path.join(newDir.uri.toFilePath(),
-      '$title.$musicFileEnding'); 
+    if (audio.container.name.toString() != musicFileEnding) {
+      // Use FFmpeg to convert to .mp3
+      String outputPath = path.join(newDir.uri.toFilePath(),
+        '$title.$musicFileEnding'); 
 
-    var command = "-i \"$filePath\" -vn -ab 128k -ar 44100 -y -codec:a libmp3lame -qscale:a 2 \"$outputPath\"";
-    await FlutterFFmpeg().execute(command)
-      .then((rc) => print("FFmpeg process exited with rc $rc"));
+      var command = "-i \"$filePath\" -vn -ab 128k -ar 44100 -y -codec:a libmp3lame -qscale:a 2 \"$outputPath\"";
+      await FlutterFFmpeg().execute(command)
+        .then((rc) => print("FFmpeg process exited with rc $rc"));
 
-    file = File(filePath);
-    if (await file.exists()) {
-      await file.delete();
+      file = File(filePath);
+      if (await file.exists()) {
+        await file.delete();
+      }
+
+      filePath = outputPath;
     }
 
+    await editMetadata(filePath.toString(), title, artist, album);
 
-
-    // await editMetadata2(filePath.toString(), title, artist, album);
-    
-    // var stream = file.openWrite(mode: FileMode.append);
-    await editMetadata2(outputPath.toString(), title, artist, album);
-    // await stream.close();
     print(">> Download finished");
   }
 
 
-  // modify the tags of the downloaded music file
-  // Future<File> editMetadata(File file, String pathFile, String title, String artist, String album) async {
-  //   MetadataEditor mde = MetadataEditor(pathFile);
-  //   return await mde.setMetadata(file, title, artist, album);
-  // }
-
-  Future<void> editMetadata2(String pathFile, String title, String artist, String album) async {
-    final mde = MetadataEditor2();
+  Future<void> editMetadata(String pathFile, String title, String artist, String album) async {
+    final mde = MetadataEditor();
     await mde.setTags(pathFile, title, artist, album);
   }
-
-  Future <void> editMetadata3(String pathFile, String title, String artist, String album) async {
-    await MetadataEditor3().setMetadata(pathFile, title, artist, album);
-    await MetadataEditor3().printMetadata(pathFile);
-  }
-
-  // Future<void> init() async {
-  //   WidgetsFlutterBinding.ensureInitialized();
-  //   await FlutterDownloader.initialize(
-  //     debug: true
-  //   );
-
-    // download
-  
-
 
   /// set new ending for music file e.g. mp3, wav or webm
   /// without the .
